finished printf

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: bajeanno <bajeanno@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/24 16:43:00 by bajeanno          #+#    #+#             */
-/*   Updated: 2022/11/27 23:26:18 by bajeanno         ###   ########lyon.fr   */
+/*   Updated: 2022/12/06 19:39:35 by bajeanno         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,12 +19,12 @@
 # include <unistd.h>
 
 int		ft_printf(const char *format, ...);
-size_t	ft_print_addr(void *ptr);
-size_t	ft_print_int(int nb);
-size_t	ft_print_unsigned(unsigned int nb);
-size_t	ft_print_hexa(int nb);
-size_t	ft_print_hexa_capslock(int nb);
-size_t	ft_putnbr_base(int nb, char *base);
-size_t	ft_putnbr_base_unsigned(size_t nb, char *base);
+int		ft_print_addr(void *ptr);
+int		ft_print_int(int nb);
+int		ft_print_unsigned(unsigned int nb);
+int		ft_print_hexa(int nb);
+int		ft_print_hexa_capslock(int nb);
+int		ft_putnbr_base(int nb, char *base);
+int		ft_putnbr_base_unsigned(size_t nb, char *base);
 
 #endif
